udemyPythonCourses/face-search-api/requirements.txt
#Python 3.11
insightface==0.7.3
onnxruntime==1.20.0
opencv-python
chromadb==0.4.22
fastapi
uvicorn
numpy<2.0
Pillow
python-multipart